+++ lib/glimmer/libui/custom_control/code_entry.rb
@@ -39,7 +39,7 @@ module Glimmer
         
         before_body do
           @syntax_highlighter = SyntaxHighlighter.new(language: language, theme: theme)
-          @font_default = {family: OS.mac? ? 'Consolas' : 'Courier', size: 13, weight: :medium, italic: :normal, stretch: :normal}
+          @font_default = {family: OS.mac? ? 'Courier New' : 'Courier', size: 13, weight: :medium, italic: :normal, stretch: :normal}
           @line = 0
           @position = 5
         end
@@ -80,12 +80,12 @@ module Glimmer
             
             # TODO cache this work if not changed between renders
             syntax_highlighter.syntax_highlighting(code).each do |token|
-              style_data = Rouge::Theme.find(theme).new.style_for(token[:token_type])
               token_text = token[:token_text].start_with?("\n") ? " #{token[:token_text]}" : token[:token_text]
               
               string(token_text) {
-                color style_data[:fg] || :black
-                background style_data[:bg] || :white
+                font @font_default.merge(italic: :italic) if token[:token_style][:italic]
+                color token[:token_style][:fg] || :black
+                background token[:token_style][:bg] || :white
               }
             end
           }